little correction : window is positionned with ints, not floats

+++ examples/dodger/src/App.cpp
@@ -12,7 +12,7 @@ void App::init()
                                                 m_config.getWindowHeight() ),
                                   m_config.getTitle(), 
                                   sf::Style::Close);
-  m_window->setPosition(sf::Vector2f(0,0) );
+  m_window->setPosition(sf::Vector2i(0,0) );
   m_stateManager.push(new StartState() );
   m_isRunning = true;
 }
